+++ flake8/__main__.py
@@ -1,0 +1,4 @@
+from flake8.main import main
+
+# python -m flake8 (with Python >= 2.7)
+main()