+++ CHANGES.rst
@@ -5,6 +5,7 @@ CHANGES
 ------------------
 
 - New option ``doctests`` to run Pyflakes checks on doctests too
+- New option ``jobs`` to launch multiple jobs in parallel
 - Fix Git and Mercurial hooks, issues #88 and #133
 - Fix crashes with Python 3.4 by upgrading dependencies
 - Fix traceback when running tests with Python 2.6